requirements.txt
# Core dependencies
python-dotenv>=1.0.0
pydantic>=2.0.0
pyyaml>=6.0

# AI/ML dependencies
openai>=1.0.0
anthropic>=0.8.0

# Media processing
pillow>=10.0.0
moviepy>=1.0.3

# Data handling
pandas>=2.0.0
numpy>=1.24.0

# API and web
requests>=2.31.0
fastapi>=0.109.0
uvicorn>=0.27.0

# Storage and state management
sqlalchemy>=2.0.0
redis>=5.0.0

# Utilities
python-dateutil>=2.8.2
validators>=0.22.0
